Add login and register methods to respective pages
