+++ frontend/src/pages/registerpage.jsx
@@ -1,13 +1,24 @@
-import React from "react";
+// NPM Packages
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
+// Project files
 import RegisterForm from "./auth/RegisterForm";
+import Auth from "../services/Auth";
 
 export default function RegisterPage() {
+    // Methods
+    async function register(registrationData) {
+        const registerSuccess = await Auth.register(registrationData);
+        if (!registerSuccess) {
+            alert("Registration failed. Check credentials and try again.");
+        }
+    }
+
     return (
         <div>
             <Link to="/mobile">
             <h1 className="iconLink"><i className="fas fa-arrow-circle-left"></i></h1></Link>
-        <RegisterForm />
+        <RegisterForm onSubmit={register} />
         </div>
-    )
+    );
 }
